New simplified Model class
Instead of remembering to handle button and timer callbacks, and dealing with properly working with the states in run, now Model does all the work, and you simply implement the stateEntered, stateLeft, and stateDo methods.

Also added a Base class for Hardware and Software Timers. Button and Timers can now take none as handlers, and gets an addHandler method to associate a handler later.