Remove i2cid
Since i2cid is defined by the sda and scl pins, let's take this out of the constructor parameters and let's determine it based on the sda/scl pins in the constructor. Will make it easier to call this class. Thought about keeping it in to be backward compatible, but may be better to just remove and fix if errors show up.