Minor change in Light and plenty of changes in Net
Added an isOn() method for Light to see if it is on or off.

for Net, the process is changed quite a bit. Now Net includes both Station (STA) and and Access Point (AP) modes - you can start it in either option. Obviously in AP mode, it will not have external web access.

Also created a simple WebServer class in Net, this will allow one to create a simple single-page web server that can display a dynamically generated HTML content.